requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
titiler.core==0.15.0
titiler.application==0.15.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
gunicorn==21.2.0
rio-tiler>=6.0.0
rasterio>=1.3.0